agents/manual_assistants/requirements.txt
docutils>=0.20.1
openai
PyYAML
requests
